Pull request feedback Also made the clipboard code more robust
